\docType{package}
\name{datalist}
\alias{datalist}
\alias{datalist-package}
\alias{package-datalist}
\title{datalist}
\description{
  An R package to convert datasets into numeric data lists
  for import into JAGS, WinBUGS and OpenBUGS.
}
\seealso{
  \code{\link{convert_data}} and
  \code{\link{generate_data}}.
}

